requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.7
django-appconf==1.0.5
django-braces==1.15.0
django-classy-tags==2.0.0
django-colorfield==0.4.5
django-easy-maps==1.1.3
django-mathfilters==1.0.0
django-model-utils==4.2.0
django-star-ratings==0.9.1
django-tinymce4-lite==1.8.0
geographiclib==1.52
geopy==2.2.0
idna==3.3
jsmin==3.0.0
numpy==1.19.5
pandas==1.3.4
Pillow==8.3.2
python-dateutil==2.8.2
pytz==2021.1
razorpay==1.2.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
swapper==1.2.0
urllib3==1.26.7
